requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
customtkinter==5.2.1
darkdetect==0.8.0
Faker==20.1.0
idna==3.6
mailtrap==2.0.1
mysql-connector-python==8.2.0
packaging==23.2
Pillow==10.1.0
protobuf==4.21.12
pyswip @ git+https://github.com/yuce/pyswip@59016e0841f56177d1b18ec08fd9b67792bd0a97
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
six==1.16.0
urllib3==2.1.0
